refactor theme editor
