+++ apps/web/src/features/settings/Appearance.tsx
@@ -4,16 +4,10 @@ import { DropdownMenu as KobalteDropdownMenu } from '@kobalte/core/dropdown-menu
 import ClipboardIcon from '@phosphor/clipboard.svg';
 import PencilIcon from '@phosphor/pencil-simple.svg';
 import PlusIcon from '@phosphor/plus.svg';
-import ShuffleIcon from '@phosphor/shuffle.svg';
 import TrashIcon from '@phosphor/trash.svg';
-import XIcon from '@phosphor/x.svg';
 import { ThemeChipPill } from '@theme/components/ThemeChipPill';
 import { ThemeChips } from '@theme/components/ThemeChips';
-import { ThemeEditorAdvanced } from '@theme/components/ThemeEditorAdvanced';
-import {
-  randomizeTheme,
-  ThemeEditorBasic,
-} from '@theme/components/ThemeEditorBasic';
+import { ThemeEditor } from '@theme/components/ThemeEditor';
 import { DEFAULT_THEMES } from '@theme/constants';
 import {
   currentThemeId,
@@ -41,7 +35,7 @@ import {
   saveTheme,
   updateTheme,
 } from '@theme/utils/themeUtils';
-import { Button, Dropdown, Layer, ToggleSwitch } from '@ui';
+import { Dropdown, Layer, ToggleSwitch } from '@ui';
 import {
   monochromeIcons,
   setMonochromeIcons,
@@ -55,8 +49,6 @@ import {
   SettingsRow,
   SettingsSection,
 } from './primitives';
-
-type EditorTab = 'basic' | 'advanced';
 
 /** Copies a theme's JSON to the clipboard (for sharing / importing elsewhere). */
 function CopyThemeButton(props: { themeId: string; name: string }) {
@@ -129,7 +121,6 @@ function ThemeSelectorRow(props: {
   onSelect: (id: string) => void;
   filter: (theme: ThemeV2) => boolean;
 }) {
-  const [editorTab, setEditorTab] = createSignal<EditorTab>('basic');
   const [editorOpen, setEditorOpen] = createSignal(false);
   // The custom theme being edited (saving writes back to it); undefined for a
   // brand-new theme.
@@ -153,7 +144,6 @@ function ThemeSelectorRow(props: {
     setEditingThemeId(undefined);
     setIsThemeSaved(false);
     setThemeName('New Theme');
-    setEditorTab('basic');
     setEditorOpen(true);
   };
 
@@ -173,7 +163,6 @@ function ThemeSelectorRow(props: {
       setIsThemeSaved(false);
       setThemeName(`${source?.name ?? 'Theme'} copy`);
     }
-    setEditorTab('basic');
     setEditorOpen(true);
   };
 
@@ -230,72 +219,15 @@ function ThemeSelectorRow(props: {
         />
       </SettingsRow>
 
-      {/* The theme editor opens inline beneath its selector as a distinct
-          active-editing block: a neutral, slightly elevated surface (one step
-          lighter than the card via Layer depth), inset + rounded so it reads as
-          a nested element. No color-forward accent. */}
+      {/* The theme editor opens inline beneath its selector; it mounts fresh
+          each time (resetting to the Basic tab). */}
       <Show when={editorOpen()}>
-        <Layer depth={3}>
-          <div class="mx-3 my-2 flex flex-col gap-3 rounded-xl border border-ink/[0.05] bg-surface px-4 py-4">
-            <div class="flex items-center gap-2">
-              <Button
-                label="Close editor"
-                onClick={closeEditor}
-                variant="ghost"
-                size="icon-sm"
-              >
-                <XIcon class="size-4" />
-              </Button>
-              <input
-                type="text"
-                value={themeName()}
-                onInput={(e) => setThemeName(e.currentTarget.value)}
-                spellcheck={false}
-                placeholder="Theme name"
-                aria-label="Theme name"
-                class="w-40 min-w-0 rounded-md border border-edge-muted bg-transparent px-2 py-1 text-xs text-ink outline-none placeholder:text-ink-extra-muted focus:border-accent"
-              />
-              <div class="flex-1" />
-              <Button
-                label="Randomize theme"
-                onPointerDown={randomizeTheme}
-                variant="ghost"
-                size="icon-sm"
-              >
-                <ShuffleIcon class="size-4" />
-              </Button>
-              <TabsInset
-                depth={3}
-                onChange={(value) => setEditorTab(value as EditorTab)}
-                list={[
-                  { value: 'basic', label: 'Basic' },
-                  { value: 'advanced', label: 'Variables' },
-                ]}
-                value={editorTab()}
-                defaultValue="basic"
-              />
-            </div>
-            <div class="relative overflow-hidden rounded-lg">
-              {/* The Basic view defines the box height; it stays mounted (just
-                  hidden) on the Variables tab so the variables list scrolls
-                  within that same height. Basic rows use dividers only; the
-                  Variables list keeps a bordered container. */}
-              <div classList={{ invisible: editorTab() !== 'basic' }}>
-                <ThemeEditorBasic />
-              </div>
-              <Show when={editorTab() === 'advanced'}>
-                <div class="absolute inset-0 overflow-y-auto rounded-lg border border-ink/[0.05] bg-surface">
-                  <ThemeEditorAdvanced />
-                </div>
-              </Show>
-            </div>
-            <div class="flex justify-end">
-              <Button variant="base" size="sm" onClick={saveCurrentTheme}>
-                Save theme
-              </Button>
-            </div>
-          </div>
-        </Layer>
+        <ThemeEditor
+          name={themeName()}
+          onNameChange={setThemeName}
+          onClose={closeEditor}
+          onSave={saveCurrentTheme}
+        />
       </Show>
     </>
   );
